add folder of automated data downloading and parsing
If you run the .bat file "run_scripts.bat":
1) You will be asked for 5 inputs to specify the treasury auction of
interest and date range.
2) A python script will then use Selenium to open a Google Chrome tab to
download the dates of relevent security auctions.
3) An R script will then download and parse the XML files
4) The final result will be a two files: one file of the treasury
auctions that were successful and all information pulled from them, and
a second file of all dates that failed to work
Note this was written using Windows OS, requires Google Chrome (using
ChromeDriver 2.35), Python 2.7, and R 3.3.3 (and even if you have all of
this compatability, be sure to download all releveant packages in each
of the python and R scripts in this folder if needed).
If you want to run this in non-automated fashion / lack compatible
software, open up "use_xml_parser.R" and follow the instructions in the
code (just need to change 5 inputs and download the dates file on your
own, all marked in the code).